pyproject.toml
[project]
name = "diagnosixtest"
version = "0.1.0"
description = ""
authors = [
    {name = "Gulnaz Artykaeva",email = "akhmetovagrm@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pydantic[email] (>=2.11.3,<3.0.0)",
    "uvicorn[standart] (>=0.34.1,<0.35.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "faker (>=37.1.0,<38.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
mypy = "^1.15.0"
black = "^25.1.0"
isort = "^6.0.1"
pylint = "^3.3.6"

# black
[tool.black]
line-length = 79
extend-exclude = "migrations/"

# isort
[tool.isort]
profile = "black"
extend_skip_glob = ["migrations"]
known_local_folder=["core", "crud", "api", "models", "schemas", "tests", "main", "utils"]
line_length = 79
lines_after_imports = 2


# mypy
[tool.mypy]
exclude = ["migrations/"]

[tool.pylint]
ignore = [".git", "__pycache__", "migrations", ".venv"]
max-line-length = 79
disable = ["C0114", "C0115", "C0116", "W1514", "W3101", "R0903", "C0103"]

# pytest
[tool.pytest.ini_options]
addopts = "-x"
testpaths = ["app/tests"]